requirements.txt
aiogram==2.19
beautifulsoup4==4.11.1
colorama==0.4.4
pyfiglet==0.8.post1
pyQiwiP2P==1.0.1
Pyrebase5
requests==2.28.1
urllib3==1.26.9
